refactor(vllm): extract decoder mechanics
